// src/design-system/tokens/build/react-native/tokens.d.ts
export const dimensionSpace0: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace25: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace50: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace75: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace100: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace125: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace150: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace200: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace250: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace300: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace400: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace500: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace550: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace600: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace700: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace800: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace900: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace1000: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpace1200: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceNegative25: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceNegative50: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceNegative100: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceNegative200: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};
export const dimensionSpaceNegative300: {
  original: string;
  number: number;
  decimal: number;
  scale: number;
};